requirements.txt
Django==1.8.8
djangorestframework==3.3.2
django-rest-swagger==0.3.4
psycopg2==2.6.1

elasticsearch<2.0.0
django-haystack==2.4.1
geopy==1.11.0
